backend/requirements.txt
# Brandy Box backend - Raspberry Pi Docker service
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.12
aiosqlite>=0.20.0
sqlalchemy>=2.0.0
pydantic-settings>=2.0.0
slowapi>=0.1.9
aiosmtplib>=3.0.0
